feat(cli): expose the session commands as a library
txcript-cli is now a library plus a three-line binary. The session
commands (list, continue, view, query) form one clap Subcommand,
SessionCommand, that the binary's Command flattens in and that can be
flattened into any other command enum; run_session drives it and takes
Options, today just the program name used in hints. The MCP server and
its rmcp/tokio dependencies sit behind a default-on `mcp` feature.

The helpers the commands are built from (find_session, find_exact,
under_dir, the --since/--until parsers, build_index, fragment) are public
so a caller can resolve ids and search the same way the commands do.
