+++ cli/Cargo.toml
@@ -10,21 +10,31 @@ repository = "https://github.com/skillsynchq/txcript"
 # from git (`cargo install --git … txcript-cli`) or a local checkout.
 publish = false
 
-# The binary keeps the crate-family name.
+# The binary keeps the crate-family name. The library exposes the same
+# commands as clap types and a `run_session` entry point.
+[lib]
+name = "txcript_cli"
+path = "src/lib.rs"
+
 [[bin]]
 name = "txcript"
 path = "src/main.rs"
+
+[features]
+default = ["mcp"]
+# `txcript mcp` and its rmcp/tokio dependencies.
+mcp = ["dep:rmcp", "dep:schemars", "dep:tokio"]
 
 [dependencies]
 chrono = "0.4.45"
 clap = { version = "4.6.1", features = ["derive", "wrap_help"] }
 clap_complete = "4.6.7"
-rmcp = { version = "2.2.0", default-features = false, features = ["macros", "server", "transport-io"] }
-schemars = "1.2.1"
+rmcp = { version = "2.2.0", default-features = false, features = ["macros", "server", "transport-io"], optional = true }
+schemars = { version = "1.2.1", optional = true }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.150"
 terminal_size = "0.4.3"
-tokio = { version = "1.48.0", features = ["io-std", "macros", "rt"] }
+tokio = { version = "1.48.0", features = ["io-std", "macros", "rt"], optional = true }
 txcript = { version = "0.9.0", path = "..", features = ["opencode", "search"] }
 # v7 for codex rollouts, v4 for every other harness — see `fresh_identity`.
 uuid = { version = "1.23.4", features = ["v4", "v7"] }
